refactor: request와 response로 DTO 패키지 분리
- DTO가 추가됨에 따라 DTO를 request와 response 패키지로 분리
- inner class이던 DetectedItem을 공통 DTO 빼서 다른 객체에서도 사용가능하도록 추출
- 변경된 패키지 구조에 맞게 import 경로 업데이트
