+++ src/main/java/store/sonyk9919/api/domain/analysis/async/AnalysisCompletionHandler.java
@@ -5,7 +5,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Component;
-import store.sonyk9919.api.domain.analysis.dto.AnalysisResponseDto;
+import store.sonyk9919.api.domain.analysis.dto.response.AnalysisResponseDto;
 import store.sonyk9919.api.domain.analysis.entity.AnalysisCache;
 import store.sonyk9919.api.domain.analysis.repository.redis.AnalysisCacheRepository;
 import store.sonyk9919.api.domain.sse.service.SseEmitterService;
